refactor(domain) : RandomBallPool의 패키지 변경
- 해당 클래스는 비지니스 로직이 포함되어 있으므로 domain 패키지로 변경
