+++ src/main/java/domain/Lotto.java
@@ -1,7 +1,5 @@
 package domain;
 
-import utils.RandomBallPool;
-
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
