+++ src/main/java/domain/RandomBallPool.java
@@ -1,4 +1,4 @@
-package utils;
+package domain;
 
 import domain.Ball;
